Add full test coverage to package
